fix(database): fix naming
